Created x and y grid
