+++ .gitignore
@@ -16,3 +16,5 @@
 npm-debug.log
 testem.log
 .DS_Store
+/typings/ember
+jsconfig.json
