+++ addon/styles/_frost-chart-svg-plot-x-grid.scss
@@ -1,0 +1,6 @@
+.frost-chart-svg-plot-x-grid {
+  line {
+    stroke: $frost-color-grey-7;
+    stroke-width: 1px;
+  }
+}
