+++ addon/styles/_frost-chart-svg-plot-y-grid.scss
@@ -1,4 +1,4 @@
-.frost-chart-svg-plot-grid {
+.frost-chart-svg-plot-y-grid {
   line {
     stroke: $frost-color-grey-7;
     stroke-width: 1px;
