+++ addon/styles/addon.scss
@@ -4,6 +4,7 @@
 @import './frost-chart-axis';
 @import './frost-chart-svg-plot-area';
 @import './frost-chart-svg-plot-bar';
-@import './frost-chart-svg-plot-grid';
 @import './frost-chart-svg-plot-line';
 @import './frost-chart-svg-plot-scatter';
+@import './frost-chart-svg-plot-x-grid';
+@import './frost-chart-svg-plot-y-grid';
