+++ addon/templates/components/frost-chart-svg-plot-grid.hbs
@@ -1,9 +1,11 @@
 {{! Template for the frost-chart-svg-plot-grid component }}
 
-{{#each _xAxisTicks as |tick|}}
-  <line x1={{tick.x}} y1={{tick.y}} x2={{tick.x}} y2=0 />
-{{/each}}
+{{frost-chart-svg-plot-x-grid
+  chartState=chartState
+  ticks=ticks
+}}
 
-{{#each _yAxisTicks as |tick|}}
-  <line x1=0 y1={{tick.y}} x2={{tick.x}} y2={{tick.y}} />
-{{/each}}
+{{frost-chart-svg-plot-y-grid
+  chartState=chartState
+  ticks=ticks
+}}
