+++ addon/templates/components/frost-chart-svg-plot-x-grid.hbs
@@ -1,0 +1,4 @@
+{{! Template for the frost-chart-svg-plot-x-grid component }}
+{{#each _ticks as |tick|}}
+  <line x1={{tick.x}} y1={{tick.y}} x2={{tick.x}} y2=0 />
+{{/each}}
