+++ addon/templates/components/frost-chart-svg-plot-y-grid.hbs
@@ -1,0 +1,4 @@
+{{! Template for the frost-chart-svg-plot-y-grid component }}
+{{#each _ticks as |tick|}}
+  <line x1=0 y1={{tick.y}} x2={{tick.x}} y2={{tick.y}} />
+{{/each}}
