+++ addon/templates/components/frost-chart-svg.hbs
@@ -21,6 +21,18 @@
         ticks=(linear-ticks 10)
       )
 
+      x-grid=(component 'frost-chart-svg-plot-x-grid'
+        chartState=chartState
+        hook=(concat hookPrefix '-xGrid')
+        ticks=(linear-ticks 10)
+      )
+
+      y-grid=(component 'frost-chart-svg-plot-y-grid'
+        chartState=chartState
+        hook=(concat hookPrefix '-yGrid')
+        ticks=(linear-ticks 10)
+      )
+
       line=(component 'frost-chart-svg-plot-line'
         chartState=chartState
         hook=(concat hookPrefix '-line')
