+++ app/components/frost-chart-svg-plot-x-grid.js
@@ -1,0 +1,4 @@
+/**
+ * Simple re-export of frost-chart-svg-plot-x-grid in the app namespace
+ */
+export {default} from 'ember-frost-chart/components/frost-chart-svg-plot-x-grid'
